[WASI] Refactor factory and algorithm. 1. The usage of factory and algorithm overlap, remove old algorithm and rename factory to algorithm, put original factory function to corresponding object file. 2.  Remove symmetric squeezeKey parmameter, use trait to get required key type. 3.  Update Context signatures because of use new algorithm.
