+++ include/host/wasi_crypto/asymmetric_common/secretkey.h
@@ -28,6 +28,10 @@ namespace AsymmetricCommon {
 
 using SkVariant = RegistedAlg::SkVariant;
 
+WasiCryptoExpect<SkVariant>
+importSk(AsymmetricCommon::Algorithm Alg, Span<const uint8_t> Encoded,
+         __wasi_secretkey_encoding_e_t Encoding) noexcept;
+
 WasiCryptoExpect<std::vector<uint8_t>>
 skExportData(const SkVariant &SkVariant,
              __wasi_secretkey_encoding_e_t Encoding) noexcept;
