+++ lib/host/wasi_crypto/symmetric/state.cpp
@@ -2,6 +2,7 @@
 // SPDX-FileCopyrightText: 2019-2022 Second State INC
 
 #include "host/wasi_crypto/symmetric/state.h"
+#include "common/errcode.h"
 #include "host/wasi_crypto/utils/error.h"
 #include "wasi_crypto/api.hpp"
 
@@ -16,7 +17,74 @@ WasiCryptoExpect<size_t> checkedAdd(size_t A, size_t B) {
                  __WASI_CRYPTO_ERRNO_OVERFLOW);
   return Res;
 }
+
+template <typename P> struct StateOpenTraits;
+template <typename StateType, typename KeyType, typename OptionsType>
+struct StateOpenTraits<WasiCryptoExpect<StateType> (*)(
+    const KeyType &, OptionalRef<const OptionsType>) noexcept> {
+  static inline constexpr bool NeedKey = true;
+};
+
+template <typename StateType, typename OptionsType>
+struct StateOpenTraits<WasiCryptoExpect<StateType> (*)(
+    OptionalRef<const OptionsType>) noexcept> {
+  static inline constexpr bool NeedKey = false;
+};
+
+template <typename T>
+inline constexpr bool NeedKey =
+    StateOpenTraits<decltype(&T::State::open)>::NeedKey;
 } // namespace
+
+WasiCryptoExpect<StateVariant>
+openState(Algorithm Alg, OptionalRef<const KeyVariant> OptKeyVariant,
+          OptionalRef<const Options> OptOptions) noexcept {
+  return std::visit(
+      [=](auto Factory) noexcept -> WasiCryptoExpect<StateVariant> {
+        using FactoryType = std::decay_t<decltype(Factory)>;
+        using KeyType = typename FactoryType::Key;
+        using StateType = typename FactoryType::State;
+        // need key
+        if constexpr (NeedKey<FactoryType>) {
+
+          ///  not have key
+          if (unlikely(!OptKeyVariant)) {
+            return WasiCryptoUnexpect(__WASI_CRYPTO_ERRNO_KEY_REQUIRED);
+          }
+
+          // have key
+          return std::visit(
+              [OptOptions](const auto &Key) -> WasiCryptoExpect<StateVariant> {
+                // key type not same
+                if constexpr (!std::is_same_v<std::decay_t<decltype(Key)>,
+                                              KeyType>) {
+                  return WasiCryptoUnexpect(__WASI_CRYPTO_ERRNO_INVALID_KEY);
+                } else {
+                  // key type same
+                  return StateType::open(Key, OptOptions)
+                      .map([](auto &&State) noexcept {
+                        return StateVariant{std::move(State)};
+                      });
+                }
+              },
+              *OptKeyVariant);
+
+        } else {
+          // not need key
+
+          // have key
+          if (unlikely(!!OptKeyVariant)) {
+            return WasiCryptoUnexpect(__WASI_CRYPTO_ERRNO_KEY_NOT_SUPPORTED);
+          }
+
+          // not have key
+          return StateType::open(OptOptions).map([](auto &&State) noexcept {
+            return StateVariant{std::move(State)};
+          });
+        }
+      },
+      Alg);
+}
 
 WasiCryptoExpect<size_t> stateOptionsGet(const StateVariant &StateVariant,
                                          std::string_view Name,
@@ -50,14 +118,33 @@ WasiCryptoExpect<Tag> stateSqueezeTag(StateVariant &StateVariant) noexcept {
                     StateVariant);
 }
 
+namespace {
+template <typename> struct GetSqueezeKeyTypeTrait;
+template <typename KeyType, typename StateType>
+struct GetSqueezeKeyTypeTrait<WasiCryptoExpect<KeyType> (
+    StateType::*)() noexcept> {
+  using Key = KeyType;
+};
+template <typename T>
+using GetSqueezeKeyType =
+    typename GetSqueezeKeyTypeTrait<decltype(&T::squeezeKey)>::Key;
+
+} // namespace
+
 WasiCryptoExpect<KeyVariant> stateSqueezeKey(StateVariant &StateVariant,
                                              Algorithm KeyAlg) noexcept {
   return std::visit(
-      [KeyAlg](auto &State) noexcept {
-        return State.squeezeKey(KeyAlg).map(
-            [](auto &&Key) { return KeyVariant{std::move(Key)}; });
+      [](auto &State, auto Alg) noexcept -> WasiCryptoExpect<KeyVariant> {
+        if constexpr (std::is_same_v<
+                          GetSqueezeKeyType<std::decay_t<decltype(State)>>,
+                          typename decltype(Alg)::Key>) {
+          return State.squeezeKey().map(
+              [](auto &&Key) { return KeyVariant{std::move(Key)}; });
+        } else {
+          return WasiCryptoUnexpect(__WASI_CRYPTO_ERRNO_UNSUPPORTED_ALGORITHM);
+        }
       },
-      StateVariant);
+      StateVariant, KeyAlg);
 }
 
 WasiCryptoExpect<size_t>
